+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API.Contracts/Project/Tasks/GetProjectByTaskResponse.cs
@@ -50,5 +50,6 @@ namespace Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks
         public DueDiligenceChecks DueDiligenceChecks { get; set; }
 
         public PreFundingAgreementCheckpointMeetingTask PreFundingAgreementCheckpointMeeting { get; set; }
+        public ReadinessToOpenMeetingTask ReadinessToOpenMeetingTask { get; set; }
     }
 }