+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API.Contracts/Project/Tasks/ProjectByTaskSummaryResponse.cs
@@ -62,6 +62,9 @@ namespace Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks
         public TaskSummaryResponse DueDiligenceChecks { get; set; }
 
         public TaskSummaryResponse PreFundingAgreementCheckpointMeeting { get; set; }
+        
+        public TaskSummaryResponse ReadinessToOpenMeeting { get; set; }
+        
         public int TaskCount { get; set; }
         
         public int CompletedTasks { get; set; }