+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects/Pages/Project/Tasks/PupilNumbersChecks/EditPupilNumbersChecksTask.cshtml.cs
@@ -7,11 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
 using System;
-using System.ComponentModel;
-using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
-using Dfe.ManageFreeSchoolProjects.Models;
-using Dfe.ManageFreeSchoolProjects.Validators;
 
 
 namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.PupilNumbersChecks
@@ -62,18 +58,12 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.PupilNumbersChecks
         {
             var project = await _getProjectService.Execute(ProjectId, TaskName.ApplicationsEvidence);
             SchoolName = project.SchoolName;
-
-            if (!ModelState.IsValid)
-            {
-                _errorService.AddErrors(ModelState.Keys, ModelState);
-                return Page();
-            }
-
+            
             try
             {
-                var request = new UpdateProjectByTaskRequest()
+                var request = new UpdateProjectByTaskRequest
                 {
-                    PupilNumbersChecks = new PupilNumbersChecksTask()
+                    PupilNumbersChecks = new PupilNumbersChecksTask
                     {
                         SchoolReceivedEnoughApplications = SchoolReceivedEnoughApplications, 
                         CapacityDataMatchesFundingAgreement = CapacityDataMatchesFundingAgreement,