+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects/TagHelpers/RadiosListItemTagHelper.cs
@@ -20,13 +20,16 @@ namespace Dfe.ManageFreeSchoolProjects.TagHelpers
 
         [HtmlAttributeName("hint")]
         public string Hint { get; set; }
+        
+        [HtmlAttributeName("name")]
+        public string Name { get; set; }
 
         public override void Process(TagHelperContext context, TagHelperOutput output)
         {
             output.TagName = "div";
             output.Attributes.SetAttribute("class", "govuk-radios__item");
             
-            output.PreContent.SetHtmlContent(RadiosListItemBuilder.BuildRadioInput(Id, Value, For));
+            output.PreContent.SetHtmlContent(RadiosListItemBuilder.BuildRadioInput(Id, Value, For, Name));
             output.PreContent.AppendHtml(RadiosListItemBuilder.BuildLabel(Id, Description));
             output.PreContent.AppendHtml(RadiosListItemBuilder.BuildHint(Id, Hint));